+++ packages/debugger/src/index.ts
@@ -1,17 +1,42 @@
-import { FlowComponent, createSignal, createComputed, onCleanup } from "solid-js"
+import { ParentComponent, createSignal, onCleanup, Accessor, createMemo } from "solid-js"
 import { isProd } from "@solid-primitives/utils"
 import { createBranch } from "@solid-primitives/rootless"
 import { getOwner } from "@shared/graph"
 import { makeBatchUpdateListener } from "./batchUpdates"
 import { createGraphRoot } from "./primitives"
-import { useLocator } from "@solid-devtools/locator"
+import { LocatorOptions, useLocator } from "@solid-devtools/locator"
 import { useExtensionAdapter } from "@solid-devtools/extension-adapter"
 
 export { makeBatchUpdateListener } from "./batchUpdates"
 
+export type { TargetIDE, TargetURLFunction } from "@solid-devtools/locator"
+
+export type DebuggerLocatorOptions = Omit<LocatorOptions, "components">
+
+export type DebuggerProps = {
+	locator?: boolean | DebuggerLocatorOptions
+}
+
+function createConsumers(): [
+	needed: Accessor<boolean>,
+	addConsumer: (consumer: Accessor<boolean>) => void,
+] {
+	const [consumers, setConsumers] = createSignal<Accessor<boolean>[]>([])
+	const enabled = createMemo<boolean>(() => consumers().some(consumer => consumer()))
+	return [enabled, consumer => setConsumers(p => [...p, consumer])]
+}
+
 let debuggerAlive = false
 
-export const Debugger: FlowComponent = props => {
+/**
+ * Debugger is a cornerstone of all solid-devtools. It analyses and tracks changes of Solid's reactive graph.
+ * Wrap your application with it to use compatable devtools.
+ *
+ * @see https://github.com/thetarnav/solid-devtools#available-devtools
+ *
+ * @param props
+ */
+export const Debugger: ParentComponent<DebuggerProps> = props => {
 	// run the debugger only on client in development env
 	if (isProd) return props.children
 
@@ -22,12 +47,12 @@ export const Debugger: FlowComponent = props => {
 
 	const root = getOwner()!
 
-	// create the graph in a separate root, so that it doesn't walk and track itself
+	// setup the debugger in a separate root, so that it doesn't walk and track itself
 	createBranch(() => {
-		const [enabled, setEnabled] = createSignal(false, { internal: true })
-		const [trackSignals, setTrackSignals] = createSignal(false, { internal: true })
-		const [trackBatchedUpdates, setTrackBatchedUpdates] = createSignal(false, { internal: true })
-		const [trackComponents, setTrackComponents] = createSignal(false, { internal: true })
+		const [enabled, addDebuggerConsumer] = createConsumers()
+		const [trackSignals, addTrackSignalsConsumer] = createConsumers()
+		const [trackBatchedUpdates, addTrackBatchedUpdatesConsumer] = createConsumers()
+		const [trackComponents, addTrackComponentsConsumer] = createConsumers()
 
 		const [{ rootId, children, components }, { forceUpdate, update }] = createGraphRoot(root, {
 			enabled,
@@ -36,24 +61,31 @@ export const Debugger: FlowComponent = props => {
 			trackComponents,
 		})
 
-		const { enabled: extensionAdapterEnabled } = useExtensionAdapter({
-			tree: {
-				id: rootId,
-				get children() {
-					return children()
+		{
+			const { enabled } = useExtensionAdapter({
+				tree: {
+					id: rootId,
+					get children() {
+						return children()
+					},
 				},
-			},
-			forceUpdate,
-			update,
-			makeBatchUpdateListener,
-		})
+				forceUpdate,
+				update,
+				makeBatchUpdateListener,
+			})
+			addDebuggerConsumer(enabled)
+			addTrackSignalsConsumer(enabled)
+			addTrackBatchedUpdatesConsumer(enabled)
+		}
 
-		const { enabled: locatorEnabled } = useLocator({ components })
-
-		createComputed(() => setEnabled(extensionAdapterEnabled() || locatorEnabled()))
-		createComputed(() => setTrackSignals(extensionAdapterEnabled))
-		createComputed(() => setTrackBatchedUpdates(extensionAdapterEnabled))
-		createComputed(() => setTrackComponents(locatorEnabled))
+		if (props.locator) {
+			const { enabled } = useLocator({
+				...(typeof props.locator === "object" ? props.locator : null),
+				components,
+			})
+			addDebuggerConsumer(enabled)
+			addTrackComponentsConsumer(enabled)
+		}
 	})
 
 	return props.children