+++ packages/locator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-devtools/locator
 
+## 0.2.1
+
+### Patch Changes
+
+- Locator module now has to be explicitely enabled with Debugger props
+  Support for different source code targets (vscode, atom, webstorm and a custom function)
+
 ## 0.2.0
 
 ### Minor Changes